perf(service): back off the provision launch poll like every other wait
provision_attempt watched for the ready sentinel or a dead child with a
hand-rolled loop that slept a flat 50 ms between looks, so a launch that
failed in the first millisecond was seen on the next tick. It now uses the
crate's shared poll_until with the same 5 ms to 50 ms backoff the readiness
wait uses: one pattern, one place to tune. The storage figures move to
vm_files/storage.rs so the file stays under its ratchet.
