+++ config/gate.toml
@@ -1090,7 +1090,7 @@ must_stay_below_lines = 3000
 "crates/capsem-service/src/ledger_routes.rs" = 2923
 "crates/capsem-service/src/main.rs" = 2112
 "crates/capsem-service/src/profile_routes.rs" = 2320
-"crates/capsem-service/src/vm_files.rs" = 1930
+"crates/capsem-service/src/vm_files.rs" = 1909
 "crates/capsem-service/src/tests/assets_registry.rs" = 1600
 "crates/capsem-service/src/tests/lifecycle.rs" = 2396
 "crates/capsem-service/src/tests/profile_routes.rs" = 2682
